requirements.txt
fastapi
uvicorn
langchain
langgraph
pydantic
langchain-community
langchain-huggingface
langchain-classic
ddgs
python-dotenv
wikipedia
streamlit